Crate wasmer_object

Object creator for Wasm Compilations.

Given a compilation result (this is, the result when calling Compiler::compile_module) this exposes functions to create an Object file for a given target.

Re-exports

pub use object;

Modules

error 🔒

module 🔒

Structs

Object

A writable object file.

Enums

ObjectError

The Object error can occur when creating an object file from a Compilation.

Functions

emit_compilation

Emit the compilation result into an existing object.

emit_data

Write data into an existing object.

emit_serialized

Emit the compilation result into an existing object.

get_object_for_target

Create an object for a given target Triple.