Trait wasmer_object::Object

pub trait Object<'data, 'file>: Sealed where
    'data: 'file, {
    type Segment: ObjectSegment<'data>;
    type SegmentIterator: Iterator;
    type Section: ObjectSection<'data>;
    type SectionIterator: Iterator;
    type Comdat: ObjectComdat<'data>;
    type ComdatIterator: Iterator;
    type Symbol: ObjectSymbol<'data>;
    type SymbolIterator: Iterator;
    type SymbolTable: ObjectSymbolTable<'data>;
    type DynamicRelocationIterator: Iterator;

    fn architecture(&self) -> Architecture;

    fn is_little_endian(&self) -> bool;

    fn is_64(&self) -> bool;

    fn kind(&self) -> ObjectKind;

    fn segments(&'file self) -> Self::SegmentIterator;

    fn section_by_name_bytes(
        &'file self, 
        section_name: &[u8]
    ) -> Option<Self::Section>;

    fn section_by_index(
        &'file self, 
        index: SectionIndex
    ) -> Result<Self::Section, Error>;

    fn sections(&'file self) -> Self::SectionIterator;

    fn comdats(&'file self) -> Self::ComdatIterator;

    fn symbol_table(&'file self) -> Option<Self::SymbolTable>;

    fn symbol_by_index(
        &'file self, 
        index: SymbolIndex
    ) -> Result<Self::Symbol, Error>;

    fn symbols(&'file self) -> Self::SymbolIterator;

    fn dynamic_symbol_table(&'file self) -> Option<Self::SymbolTable>;

    fn dynamic_symbols(&'file self) -> Self::SymbolIterator;

    fn dynamic_relocations(
        &'file self
    ) -> Option<Self::DynamicRelocationIterator>;

    fn imports(&self) -> Result<Vec<Import<'data>, Global>, Error>;

    fn exports(&self) -> Result<Vec<Export<'data>, Global>, Error>;

    fn has_debug_symbols(&self) -> bool;

    fn relative_address_base(&'file self) -> u64;

    fn entry(&'file self) -> u64;

    fn flags(&self) -> FileFlags;

    fn endianness(&self) -> Endianness { ... }

    fn section_by_name(&'file self, section_name: &str) -> Option<Self::Section> { ... }

    fn symbol_map(&'file self) -> SymbolMap<SymbolMapName<'data>> { ... }

    fn object_map(&'file self) -> ObjectMap<'data> { ... }

    fn mach_uuid(&self) -> Result<Option<[u8; 16]>, Error> { ... }

    fn build_id(&self) -> Result<Option<&'data [u8]>, Error> { ... }

    fn gnu_debuglink(&self) -> Result<Option<(&'data [u8], u32)>, Error> { ... }

    fn gnu_debugaltlink(
        &self
    ) -> Result<Option<(&'data [u8], &'data [u8])>, Error> { ... }

    fn pdb_info(&self) -> Result<Option<CodeView<'_>>, Error> { ... }

}

An object file.

Associated Types

type Segment: ObjectSegment<'data>

A segment in the object file.

type SegmentIterator: Iterator

An iterator over the segments in the object file.

type Section: ObjectSection<'data>

A section in the object file.

type SectionIterator: Iterator

An iterator over the sections in the object file.

type Comdat: ObjectComdat<'data>

A COMDAT section group in the object file.

type ComdatIterator: Iterator

An iterator over the COMDAT section groups in the object file.

type Symbol: ObjectSymbol<'data>

A symbol in the object file.

type SymbolIterator: Iterator

An iterator over symbols in the object file.

type SymbolTable: ObjectSymbolTable<'data>

A symbol table in the object file.

type DynamicRelocationIterator: Iterator

An iterator over dynamic relocations in the file.

The first field in the item tuple is the address that the relocation applies to.

Required methods

fn architecture(&self) -> Architecture

Get the architecture type of the file.

fn is_little_endian(&self) -> bool

Return true if the file is little endian, false if it is big endian.

fn is_64(&self) -> bool

Return true if the file can contain 64-bit addresses.

fn kind(&self) -> ObjectKind

Return the kind of this object.

fn segments(&'file self) -> Self::SegmentIterator

Get an iterator over the segments in the file.

fn section_by_name_bytes(
&'file self,
section_name: &[u8]
) -> Option<Self::Section>

Like Self::section_by_name, but allows names that are not UTF-8.

fn section_by_index(
&'file self,
index: SectionIndex
) -> Result<Self::Section, Error>

Get the section at the given index.

The meaning of the index depends on the object file.

For some object files, this requires iterating through all sections.

Returns an error if the index is invalid.

fn sections(&'file self) -> Self::SectionIterator

Get an iterator over the sections in the file.

fn comdats(&'file self) -> Self::ComdatIterator

Get an iterator over the COMDAT section groups in the file.

fn symbol_table(&'file self) -> Option<Self::SymbolTable>

Get the symbol table, if any.

fn symbol_by_index(
&'file self,
index: SymbolIndex
) -> Result<Self::Symbol, Error>

Get the debugging symbol at the given index.

The meaning of the index depends on the object file.

Returns an error if the index is invalid.

fn symbols(&'file self) -> Self::SymbolIterator

Get an iterator over the debugging symbols in the file.

This may skip over symbols that are malformed or unsupported.

For Mach-O files, this does not include STAB entries.

fn dynamic_symbol_table(&'file self) -> Option<Self::SymbolTable>

Get the dynamic linking symbol table, if any.

Only ELF has a separate dynamic linking symbol table.

fn dynamic_symbols(&'file self) -> Self::SymbolIterator

Get an iterator over the dynamic linking symbols in the file.

This may skip over symbols that are malformed or unsupported.

Only ELF has separate dynamic linking symbols. Other file formats will return an empty iterator.

fn dynamic_relocations(&'file self) -> Option<Self::DynamicRelocationIterator>

Get the dynamic relocations for this file.

Symbol indices in these relocations refer to the dynamic symbol table.

Only ELF has dynamic relocations.

fn imports(&self) -> Result<Vec<Import<'data>, Global>, Error>

Get the imported symbols.

fn exports(&self) -> Result<Vec<Export<'data>, Global>, Error>

Get the exported symbols that expose both a name and an address.

Some file formats may provide other kinds of symbols, that can be retrieved using the lower-level API.

fn has_debug_symbols(&self) -> bool

Return true if the file contains debug information sections, false if not.

fn relative_address_base(&'file self) -> u64

Get the base address used for relative virtual addresses.

Currently this is only non-zero for PE.

fn entry(&'file self) -> u64

Get the virtual address of the entry point of the binary

fn flags(&self) -> FileFlags

File flags that are specific to each file format.

Provided methods

fn endianness(&self) -> Endianness

Get the endianness of the file.

fn section_by_name(&'file self, section_name: &str) -> Option<Self::Section>

Get the section named section_name, if such a section exists.

If section_name starts with a ‘.’ then it is treated as a system section name, and is compared using the conventions specific to the object file format. This includes:

if “.debug_str_offsets” is requested for a Mach-O object file, then the actual section name that is searched for is “__debug_str_offs”.

if “.debug_info” is requested for an ELF object file, then “.zdebug_info” may be returned (and similarly for other debug sections).

For some object files, multiple segments may contain sections with the same name. In this case, the first matching section will be used.

This method skips over sections with invalid names.

fn symbol_map(&'file self) -> SymbolMap<SymbolMapName<'data>>

Construct a map from addresses to symbol names.

The map will only contain defined text and data symbols. The dynamic symbol table will only be used if there are no debugging symbols.

fn object_map(&'file self) -> ObjectMap<'data>

Construct a map from addresses to symbol names and object file names.

This is derived from Mach-O STAB entries.

fn mach_uuid(&self) -> Result<Option<[u8; 16]>, Error>

The UUID from a Mach-O LC_UUID load command.

fn build_id(&self) -> Result<Option<&'data [u8]>, Error>

The build ID from an ELF NT_GNU_BUILD_ID note.

fn gnu_debuglink(&self) -> Result<Option<(&'data [u8], u32)>, Error>

The filename and CRC from a .gnu_debuglink section.

fn gnu_debugaltlink(&self) -> Result<Option<(&'data [u8], &'data [u8])>, Error>

The filename and build ID from a .gnu_debugaltlink section.

fn pdb_info(&self) -> Result<Option<CodeView<'_>>, Error>

The filename and GUID from the PE CodeView section